+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
     - Issue 79 - Generate data for use by benchmarks
     - Issue 88 - There are no CloudWatch logs for the API Gateway
+    - Issue 13 - Add SWORD version from shp.xml to DB entries
 ### Changed
     - Issue 59 - Using GeoDataFrame to work with data returned from db
     - Issue 75 - Update log messaging format